requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
Werkzeug==2.3.7
WTForms==3.1.2
SQLAlchemy==2.0.26
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
gunicorn==21.2.0
google-generativeai==0.3.1
requests==2.31.0
Jinja2>=3.0.0
email-validator==2.1.0.post1
numpy==1.26.0
scipy==1.11.4
spacy==3.5.0
spotipy==2.23.0
matplotlib==3.8.0
pillow==10.1.0
scikit-learn==1.3.2
alembic==1.13.1
greenlet==3.0.1
six==1.16.0
urllib3==2.1.0
psutil==5.9.6
structlog==23.2.0
python-json-logger==2.0.7
retrying==1.3.4
tenacity==8.2.3
packaging>=21.0
schedule==1.2.1
pytest>=7.0.0